sprints: rename milestones->sprints CLI; move sprint ops to tedi_ops
- config-free sprint operations (expand_and_refresh, sync_blocker_changes +
  issue-loading helpers) move to tedi_ops/sprints.rs; the bin keeps config,
  CLI dispatch, healthcheck, and GitHub milestone-metadata calls
- CLI subcommand milestones -> sprints (user-facing); shell_init tmh/tmg/tme
  aliases updated; integration suite milestones.rs -> sprints.rs

Deferred per plan's own YAGNI guidance: TaskView and Bottleneck primitives +
core/slices.rs. TaskView differs from Milestone only by admitting milestone
refs, which no command uses yet; Bottleneck's convention is explicitly to be
decided when its first command lands. Sprint classification (title-as-Timeframe,
KEY_MILESTONES series) already lives in the command; a thin core wrapper adds
no behavior today, so it waits for a consumer.
