+++ tedi_ops/tests/integration/main.rs
@@ -6,8 +6,8 @@ pub use common::*;
 mod blocker_integrated;
 mod file_naming;
 mod issue_preservation;
-mod milestones;
 mod reset_conflict;
+mod sprints;
 mod sync;
 mod touch;
 mod touch_parent;
